requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-google-genai
google-generativeai
sentence-transformers
faiss-cpu
PyPDF2
python-dotenv
